+++ cmd/server/main.go
@@ -3,19 +3,19 @@ package main
 import (
 	"context"
 	"fmt"
+	"github.com/kosimovsky/tricMe/internal/server"
 	"log"
 	"net/http"
 	"os"
 	"os/signal"
 	"syscall"
-	"time"
 )
 
 const serverPort = 8080
 
 func main() {
 
-	srv := new(Server)
+	srv := server.NewServer()
 
 	go func() {
 		mux := http.DefaultServeMux
@@ -34,23 +34,4 @@ func main() {
 	if err := srv.Shutdown(context.Background()); err != nil {
 		log.Fatalf("error occured while server shutting down : %s", err.Error())
 	}
-}
-
-type Server struct {
-	httpServer *http.Server
-}
-
-func (s *Server) Run(port string, handler http.Handler) error {
-	s.httpServer = &http.Server{
-		Addr:           ":" + port,
-		Handler:        handler,
-		MaxHeaderBytes: 1 << 20,
-		ReadTimeout:    10 * time.Second,
-		WriteTimeout:   10 * time.Second,
-	}
-	return s.httpServer.ListenAndServe()
-}
-
-func (s *Server) Shutdown(ctx context.Context) error {
-	return s.httpServer.Shutdown(ctx)
 }
