+++ internal/models.go
@@ -1,0 +1,3 @@
+package internal
+
+type Gauge = float64
